# openapi.yaml
openapi: 3.0.0
info:
  title: Bank Rates API
  description: API for accessing information about banks.
  version: 1.0.0
servers:
  - url: http://localhost:8091
paths:
  /rates:
    post:
      summary: Retrieve bank rates
      description: Retrieve bank rates for specified bank IDs, currency ID, and date.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bankIds:
                  type: array
                  items:
                    type: integer
                  description: List of bank IDs
                currencyId:
                  type: integer
                  description: Currency ID
                date:
                  type: string
                  format: date
                  description: Date in YYYY-MM-DD format. If not provided, current date is used.
      responses:
        '200':
          description: Bank rates retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BankRate'
        '500':
          description: Internal server error
  /banks:
    get:
      summary: Retrieve banks
      description: |
        Retrieve a list of banks.
      responses:
        '200':
          description: A list of banks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bank'
        '500':
          description: Internal server error
  /currencies:
    get:
      summary: Retrieve Currencies
      description: |
        Retrieve a list of currencies.
      responses:
        '200':
          description: A list of currencies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Currency'
        '500':
          description: Internal server error
components:
  schemas:
    Bank:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: The ID of the bank.
        name:
          type: string
          description: The name of the bank.
    Currency:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: The ID of the currency.
        name:
          type: string
          description: The name of the currency.
        symbol:
          type: string
          description: The symbol of the currency.
    